Refactor ogbg Jax workload
